+++ bolt-spammer-helder/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod constants;
 /// The on-chain registry view for querying active Bolt sidecars
-pub mod onchain_registry;
-use onchain_registry::BoltRegistry;
+pub mod contract;
+use contract::BoltRegistry;
 
 pub mod utils;
